Data Binding and Fragments Implementations
Two fragments are used in a activity like two sides of a coin where coin is the activity. Then going to another fragment from a fragment by swiping is disabled. For going to a fragment from another fragments two button is used. While going to 2nd fragment, the 2nd fragment always refresh itself but not the 1st fragment. To come back to 1st fragment from 2nd fragment by back button is implmented. That is two is working like they are two activities but they are not actually.

Data binding is enabled. So findViewById(...) is not used to bind a view declared in XML.